Fix i32 casts in generic_heap_tests.rs - casts are necessary

+++ tests/generic_heap_tests.rs
@@ -90,12 +90,12 @@ fn test_multiple_decrease_keys<H: Heap<i32, i32>>() {
 
     // Insert 20 elements with high priorities
     for i in 0..20 {
-        handles.push(heap.push((i + 1) * 100, i));
+        handles.push(heap.push(((i + 1) * 100) as i32, i as i32));
     }
 
     // Decrease all keys to be much smaller
     for (i, handle) in handles.iter().enumerate() {
-        heap.decrease_key(handle, i);
+        heap.decrease_key(handle, i as i32);
     }
 
     // Verify heap property maintained
@@ -202,7 +202,7 @@ fn test_stress_operations<H: Heap<i32, i32>>() {
 
     // Random decrease_key operations
     for i in (0..100).step_by(3) {
-        heap.decrease_key(&handles[i], i * 2 - 1);
+        heap.decrease_key(&handles[i], (i * 2 - 1) as i32);
     }
 
     // Pop some elements
